Use a DummyAPI implementation
